Fix bug caused by depending on drop=TRUE to work
A bug in insight caused insight::get_predictors() to return the model
frame (a data frame), on which drop defaults to TRUE, and so the column
selection was a vector. With the bug fixed, it now returns the original
object, which could be a tibble, and tibbles default to drop=FALSE. Use
the correct subsetting to always get a vector from the column.

See https://github.com/easystats/insight/pull/761 for the insight bug.