Capture the parent frame of population() for eval later
It seems reasonable to expect population() to be lexically scoped, so
that expressions inside (such as response expressions, error scales, and
sizes) are evaluated with access to both (a) the relevant simulated data
and (b) variables in the environment where the population was defined.
This allows, for instance, for a simulated population's parameters to
depend on local variables, which might be useful in simulation studies.

To enable this, capture the calling environment in each population()
object. When evaluating expressions during sampling, use this as the
enclosing environment, so symbols are looked up first in the simulated
data frame and second in the population's lexical environment.