Detect if model is fit without `data =` argument
If the user uses `$` in the model formula to refer to data frames in the
local environment, our simulation functions will not work correctly:
they call `update(fit, data = ...)` with the simulated data, but the
refit model doesn't use the `data` argument. And so all "simulations"
will actually just be fits to the original data.

We can automatically detect a missing `data =` argument and complain. Of
course, the user can still provide it and use `$`, but at least this
reduces the odds of a footgun.

Closes #4.