Make parametric_boot_distribution() work for null hypothesis tests
For tests where `fit` has fewer predictors than `alternative_fit`, the
previous version failed because the simulated data frame did not contain
the additional predictors needed to refit `alternative_fit`. This
precluded simulation-based inference for predictors.

We could use the data from `alternative_fit`, but we'd have to ensure
that both models were fit to the same data, or the results would be
absurd. Instead of attempting to verify that, make it the user's
responsibility to provide a data frame when needed.