Switch to prefer predictor distributions provided as functions
I used strings because printing out a nice representation of the
population is easier when the predictor distributions are given as
strings. But this can cause scoping issues when do.call() needs to find
the function; this is a particular issue when the user defines their own
predictor distribution function. Switch to using the function directly,
and get a printable representation of it with quote(substitute()) for
display.

Strings are still accepted, since do.call() will use them, but may hit
scoping issues occasionally.