Ensure partial_residuals() can be calculated with offsets
In a model with an offset term, the predictors do not include the
offset, so partial_residuals() previously failed. If there is an offset,
we must include it when making predictions. Fix the offset predictor at
its mean value. (Other predictors are fixed at 0, but offsets are often
log-transformed, so 0 would be unwise.)