Make binning heuristic depend on number of non-NA rows, not total
bin_by_quantile() and bin_by_interval() have a default heuristic for the
number of bins, based on the total number of rows in the data. But if
the variable in question has a high rate of missingness, this can choose
too many bins. Base the heuristic on the number of non-NA values of the
variable of interest, rather than the total number of rows.