Switch augment_longer() to use predictors, not regressors
In a model such as y ~ x + I(x^2), using regressors results in separate
rows for x and x^2. This is not very useful when faceting residual
plots: a plot of residuals vs x is the same as a plot of residuals vs
x^2, just with an axis transformation. Instead, use the unique
predictors.

Some vignettes already claimed it used predictors, not regressors, so
adjust everything to be consistent and add examples and tests.